+++ src/main/java/net/geforcemods/securitycraft/datagen/ItemModelGenerator.java
@@ -6,6 +6,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import net.geforcemods.securitycraft.ClientHandler;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.api.IReinforcedBlock;
@@ -14,6 +15,7 @@ import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedButtonBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedPistonBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedStainedGlassPaneBlock;
 import net.geforcemods.securitycraft.blocks.reinforced.ReinforcedWallBlock;
+import net.geforcemods.securitycraft.items.CodebreakerItem;
 import net.minecraft.block.Block;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.item.Item;
@@ -71,10 +73,16 @@ public class ItemModelGenerator extends ItemModelProvider {
 				SCContent.UNIVERSAL_BLOCK_REINFORCER_LVL_2,
 				SCContent.UNIVERSAL_BLOCK_REINFORCER_LVL_3,
 				SCContent.UNIVERSAL_KEY_CHANGER);
+		List<RegistryObject<Item>> linkingStateItems = Arrays.asList(
+				SCContent.CAMERA_MONITOR,
+				SCContent.MINE_REMOTE_ACCESS_TOOL,
+				SCContent.SENTRY_REMOTE_ACCESS_TOOL,
+				SCContent.SONIC_SECURITY_SYSTEM_ITEM);
 
 		singleTextureItems.removeAll(Arrays.asList(
 				SCContent.ANCIENT_DEBRIS_MINE_ITEM,
 				SCContent.BRIEFCASE,
+				SCContent.CODEBREAKER,
 				SCContent.DISPLAY_CASE,
 				SCContent.KEYCARD_HOLDER,
 				SCContent.KEYPAD_CHEST_ITEM,
@@ -88,6 +96,7 @@ public class ItemModelGenerator extends ItemModelProvider {
 				SCContent.UNIVERSAL_OWNER_CHANGER,
 				SCContent.WIRE_CUTTERS));
 		singleTextureItems.removeAll(handheldItems);
+		singleTextureItems.removeAll(linkingStateItems);
 		//@formatter:on
 
 		for (RegistryObject<Item> obj : singleTextureItems) {
@@ -110,6 +119,8 @@ public class ItemModelGenerator extends ItemModelProvider {
 		.element().from(0, 0, 0).to(16, 16, 16).face(Direction.UP).cullface(Direction.UP).texture("#overlay").end().end();
 		//@formatter:on
 
+		linkingStateItems.forEach(this::linkingStateItem);
+		codebreaker();
 		blockMine(SCContent.ANCIENT_DEBRIS_MINE.get(), mcLoc(BLOCK_FOLDER + "/ancient_debris_side"), mcLoc(BLOCK_FOLDER + "/ancient_debris_side"), mcLoc(BLOCK_FOLDER + "/ancient_debris_top"));
 		blockMine(SCContent.FURNACE_MINE.get(), mcLoc(BLOCK_FOLDER + "/furnace_side"), mcLoc(BLOCK_FOLDER + "/furnace_front"), mcLoc(BLOCK_FOLDER + "/furnace_top"));
 		blockMine(SCContent.SMOKER_MINE.get(), mcLoc(BLOCK_FOLDER + "/smoker_side"), mcLoc(BLOCK_FOLDER + "/smoker_front"), mcLoc(BLOCK_FOLDER + "/smoker_top"));
@@ -130,6 +141,42 @@ public class ItemModelGenerator extends ItemModelProvider {
 		reinforcedWallInventory(SCContent.REINFORCED_RED_NETHER_BRICK_WALL.get(), "red_nether_bricks");
 		reinforcedWallInventory(SCContent.REINFORCED_END_STONE_BRICK_WALL.get(), "end_stone_bricks");
 		reinforcedWallInventory(SCContent.REINFORCED_POLISHED_BLACKSTONE_BRICK_WALL.get(), "polished_blackstone_bricks");
+	}
+
+	public void linkingStateItem(RegistryObject<Item> item) {
+		String hasLinkedPath = item.getId().getPath();
+		String defaultPath = hasLinkedPath + "_idle";
+		String notLinkedPath = hasLinkedPath + "_not_linked";
+		String linkedPath = hasLinkedPath + "_linked";
+
+		//@formatter:off
+		simpleItem(defaultPath, "item/generated");
+		simpleItem(notLinkedPath, "item/generated");
+		simpleItem(linkedPath, "item/generated");
+		simpleItem(hasLinkedPath, "item/generated")
+			.override().predicate(ClientHandler.LINKING_STATE_PROPERTY, ClientHandler.EMPTY_STATE).model(new UncheckedModelFile(modItem(defaultPath))).end()
+			.override().predicate(ClientHandler.LINKING_STATE_PROPERTY, ClientHandler.UNKNOWN_STATE).model(new UncheckedModelFile(modItem(hasLinkedPath))).end()
+			.override().predicate(ClientHandler.LINKING_STATE_PROPERTY, ClientHandler.NOT_LINKED_STATE).model(new UncheckedModelFile(modItem(notLinkedPath))).end()
+			.override().predicate(ClientHandler.LINKING_STATE_PROPERTY, ClientHandler.LINKED_STATE).model(new UncheckedModelFile(modItem(linkedPath))).end();
+		//@formatter:on
+	}
+
+	public void codebreaker() {
+		String defaultPath = SCContent.CODEBREAKER.getId().getPath();
+		String decodingPath = defaultPath + "_decoding";
+		String failurePath = defaultPath + "_failure";
+		String successPath = defaultPath + "_success";
+
+		//@formatter:off
+		simpleItem(decodingPath, "item/generated");
+		simpleItem(failurePath, "item/generated");
+		simpleItem(successPath, "item/generated");
+		simpleItem(defaultPath, "item/generated")
+			.override().predicate(CodebreakerItem.STATE_PROPERTY, 0.0F).model(new UncheckedModelFile(modItem(defaultPath))).end()
+			.override().predicate(CodebreakerItem.STATE_PROPERTY, 0.25F).model(new UncheckedModelFile(modItem(decodingPath))).end()
+			.override().predicate(CodebreakerItem.STATE_PROPERTY, 0.5F).model(new UncheckedModelFile(modItem(failurePath))).end()
+			.override().predicate(CodebreakerItem.STATE_PROPERTY, 0.75F).model(new UncheckedModelFile(modItem(successPath))).end();
+		//@formatter:on
 	}
 
 	public ItemModelBuilder simpleItem(Item item, String parent) {