+++ src/main/java/net/geforcemods/securitycraft/items/SonicSecuritySystemItem.java
@@ -200,7 +200,7 @@ public class SonicSecuritySystemItem extends BlockItem {
 	 * @return true if the tag contains at least one position, false otherwise
 	 */
 	public static boolean hasLinkedBlock(CompoundNBT tag) {
-		if (!tag.contains("LinkedBlocks"))
+		if (tag == null || !tag.contains("LinkedBlocks"))
 			return false;
 
 		return !tag.getList("LinkedBlocks", Constants.NBT.TAG_COMPOUND).isEmpty();