+++ src/main/java/net/geforcemods/securitycraft/screen/CameraMonitorScreen.java
@@ -90,8 +90,8 @@ public class CameraMonitorScreen extends Screen {
 
 		for (int i = 0; i < 10; i++) {
 			CameraButton button = cameraButtons[i];
-			int camID = (button.camId + ((page - 1) * 10));
-			List<GlobalPos> views = cameraMonitor.getCameraPositions(nbtTag);
+			int camID = button.camId + ((page - 1) * 10);
+			List<GlobalPos> views = CameraMonitorItem.getCameraPositions(nbtTag);
 			GlobalPos view = views.get(camID - 1);
 
 			button.setMessage(button.getMessage().plainCopy().append(new StringTextComponent("" + camID)));
@@ -126,10 +126,10 @@ public class CameraMonitorScreen extends Screen {
 		if (page == 1)
 			prevPageButton.active = false;
 
-		if (page == 3 || cameraMonitor.getCameraPositions(nbtTag).size() < (page * 10) + 1)
+		if (page == 3 || CameraMonitorItem.getCameraPositions(nbtTag).size() < (page * 10) + 1)
 			nextPageButton.active = false;
 
-		for (int i = cameraMonitor.getCameraPositions(nbtTag).size() + 1; i <= (page * 10); i++) {
+		for (int i = CameraMonitorItem.getCameraPositions(nbtTag).size() + 1; i <= (page * 10); i++) {
 			cameraButtons[(i - 1) - ((page - 1) * 10)].active = false;
 		}
 	}
@@ -162,7 +162,7 @@ public class CameraMonitorScreen extends Screen {
 
 	protected void cameraButtonClicked(Button button) {
 		int camID = ((CameraButton) button).camId + (page - 1) * 10;
-		BlockPos cameraPos = cameraMonitor.getCameraPositions(nbtTag).get(camID - 1).pos();
+		BlockPos cameraPos = CameraMonitorItem.getCameraPositions(nbtTag).get(camID - 1).pos();
 		TileEntity te = minecraft.level.getBlockEntity(cameraPos);
 
 		if (te instanceof SecurityCameraBlockEntity && ((SecurityCameraBlockEntity) te).isDisabled()) {
@@ -178,7 +178,7 @@ public class CameraMonitorScreen extends Screen {
 		int camID = ((CameraButton) button).camId + (page - 1) * 10;
 
 		SecurityCraft.channel.sendToServer(new RemoveCameraTag(camID));
-		nbtTag.remove(CameraMonitorItem.getTagNameFromPosition(nbtTag, cameraMonitor.getCameraPositions(nbtTag).get(camID - 1)));
+		nbtTag.remove(CameraMonitorItem.getTagNameFromPosition(nbtTag, CameraMonitorItem.getCameraPositions(nbtTag).get(camID - 1)));
 		button.active = false;
 		cameraButtons[(camID - 1) % 10].active = false;
 	}